feat: extract separable feature modules into a dockview-modules package
Introduces `dockview-modules`, a sibling package (depends on dockview-core)
that houses four feature modules previously built into core: TabGroupChips,
ContextMenu, AdvancedDnD and Accessibility.

- core gains a global module registry — registerModules / getRegisteredModules
  / clearRegisteredModules; DockviewComponent's default module set becomes
  `[...AllModules, ...getRegisteredModules()]`
- core keeps only the interfaces for the moved modules (moduleContracts +
  ServiceCollection slots + option types); the four implementations move to
  dockview-modules. allModules.ts trimmed to the seven built-ins
- the few internals the modules need (defineModule, the contracts,
  resolveMessages, findRelativeZIndexParent, IDragGhostSpec, LiveRegionModule,
  DockviewGroupDropLocation) are now exported from dockview-core's main index
- `dockview` registers the modules at import time, so the batteries-included
  entry keeps the full feature set; `dockview-core` alone ships the built-ins
- module-behaviour tests relocated to dockview-modules; framework package test
  configs resolve dockview-modules
