+++ packages/dockview-core/src/__tests__/dockview/dockviewComponent.spec.ts
@@ -5279,92 +5279,6 @@ describe('dockviewComponent', () => {
         expect(events.length).toBe(5);
     });
 
-    test('that dragging a tab triggers onWillDragPanel', () => {
-        const container = document.createElement('div');
-
-        const dockview = new DockviewComponent(container, {
-            createComponent(options) {
-                switch (options.name) {
-                    case 'default':
-                        return new PanelContentPartTest(
-                            options.id,
-                            options.name
-                        );
-                    default:
-                        throw new Error(`unsupported`);
-                }
-            },
-        });
-
-        dockview.layout(1000, 500);
-
-        dockview.addPanel({
-            id: 'panel_1',
-            component: 'default',
-        });
-
-        const tabDragEvents: TabDragEvent[] = [];
-        const groupDragEvents: GroupDragEvent[] = [];
-
-        dockview.onWillDragPanel((event) => {
-            tabDragEvents.push(event);
-        });
-        dockview.onWillDragGroup((event) => {
-            groupDragEvents.push(event);
-        });
-
-        const el = dockview.element.querySelector('.dv-tab')!;
-        expect(el).toBeTruthy();
-
-        fireEvent.dragStart(el);
-
-        expect(tabDragEvents.length).toBe(1);
-        expect(groupDragEvents.length).toBe(0);
-    });
-
-    test('that dragging a group triggers onWillDragGroup', () => {
-        const container = document.createElement('div');
-
-        const dockview = new DockviewComponent(container, {
-            createComponent(options) {
-                switch (options.name) {
-                    case 'default':
-                        return new PanelContentPartTest(
-                            options.id,
-                            options.name
-                        );
-                    default:
-                        throw new Error(`unsupported`);
-                }
-            },
-        });
-
-        dockview.layout(1000, 500);
-
-        dockview.addPanel({
-            id: 'panel_1',
-            component: 'default',
-        });
-
-        const tabDragEvents: TabDragEvent[] = [];
-        const groupDragEvents: GroupDragEvent[] = [];
-
-        dockview.onWillDragPanel((event) => {
-            tabDragEvents.push(event);
-        });
-        dockview.onWillDragGroup((event) => {
-            groupDragEvents.push(event);
-        });
-
-        const el = dockview.element.querySelector('.dv-void-container')!;
-        expect(el).toBeTruthy();
-
-        fireEvent.dragStart(el);
-
-        expect(tabDragEvents.length).toBe(0);
-        expect(groupDragEvents.length).toBe(1);
-    });
-
     test('corrupt layout: bad inline view', () => {
         const container = document.createElement('div');
 
@@ -9178,70 +9092,6 @@ describe('dockviewComponent', () => {
             jest.runAllTimers();
             expect(didLayoutChangeHandler).toHaveBeenCalledTimes(2);
         });
-
-        test('when tab groups are created, mutated, collapsed, or destroyed', () => {
-            const panel1 = dockview.addPanel({
-                id: 'panel_1',
-                component: 'default',
-            });
-            const panel2 = dockview.addPanel({
-                id: 'panel_2',
-                component: 'default',
-                position: { referenceGroup: panel1.group },
-            });
-            jest.runAllTimers();
-
-            const didLayoutChangeHandler = jest.fn();
-            dockview.onDidLayoutChange(didLayoutChangeHandler);
-
-            // create tab group
-            const tg = dockview.api.createTabGroup({
-                groupId: panel1.group.id,
-                label: 'My Group',
-                color: 'red',
-            });
-            jest.runAllTimers();
-            expect(didLayoutChangeHandler).toHaveBeenCalledTimes(1);
-
-            // add panels to tab group
-            dockview.api.addPanelToTabGroup({
-                groupId: panel1.group.id,
-                tabGroupId: tg.id,
-                panelId: panel1.id,
-            });
-            jest.runAllTimers();
-            expect(didLayoutChangeHandler).toHaveBeenCalledTimes(2);
-
-            dockview.api.addPanelToTabGroup({
-                groupId: panel1.group.id,
-                tabGroupId: tg.id,
-                panelId: panel2.id,
-            });
-            jest.runAllTimers();
-            expect(didLayoutChangeHandler).toHaveBeenCalledTimes(3);
-
-            // collapse tab group
-            tg.collapse();
-            jest.runAllTimers();
-            expect(didLayoutChangeHandler).toHaveBeenCalledTimes(4);
-
-            // remove a panel from the tab group (group still has panel2 so
-            // it is not auto-destroyed)
-            dockview.api.removePanelFromTabGroup({
-                groupId: panel1.group.id,
-                panelId: panel1.id,
-            });
-            jest.runAllTimers();
-            expect(didLayoutChangeHandler).toHaveBeenCalledTimes(5);
-
-            // explicitly dissolve the tab group
-            dockview.api.dissolveTabGroup({
-                groupId: panel1.group.id,
-                tabGroupId: tg.id,
-            });
-            jest.runAllTimers();
-            expect(didLayoutChangeHandler).toHaveBeenCalledTimes(6);
-        });
     });
 
     describe('panel visibility', () => {
@@ -11151,231 +11001,6 @@ describe('dockviewComponent', () => {
             });
             expect(restored[0].collapsed).toBe(true);
         });
-
-        test('fromJSON with tab groups fires correct events', () => {
-            dockview.layout(1000, 1000);
-
-            // Set up panels and tab groups first
-            const panel1 = dockview.addPanel({
-                id: 'panel1',
-                component: 'default',
-            });
-            dockview.addPanel({
-                id: 'panel2',
-                component: 'default',
-            });
-
-            const groupId = panel1.group.id;
-            const tg = dockview.api.createTabGroup({
-                groupId,
-                label: 'Events',
-                color: 'green',
-            });
-            dockview.api.addPanelToTabGroup({
-                groupId,
-                tabGroupId: tg.id,
-                panelId: 'panel1',
-            });
-            dockview.api.addPanelToTabGroup({
-                groupId,
-                tabGroupId: tg.id,
-                panelId: 'panel2',
-            });
-
-            const state = dockview.toJSON();
-
-            // Now restore from JSON and track events on the new groups
-            dockview.fromJSON(state);
-
-            const restoredGroup = dockview.api.panels[0].group;
-
-            const created: string[] = [];
-            const destroyed: string[] = [];
-            const panelsAdded: { tgId: string; panelId: string }[] = [];
-            const panelsRemoved: { tgId: string; panelId: string }[] = [];
-            const changes: string[] = [];
-
-            const disposable = new CompositeDisposable(
-                dockview.api.onDidCreateTabGroup((e) =>
-                    created.push(e.tabGroup.id)
-                ),
-                dockview.api.onDidDestroyTabGroup((e) =>
-                    destroyed.push(e.tabGroup.id)
-                ),
-                dockview.api.onDidAddPanelToTabGroup((e) =>
-                    panelsAdded.push({
-                        tgId: e.tabGroup.id,
-                        panelId: e.panelId,
-                    })
-                ),
-                dockview.api.onDidRemovePanelFromTabGroup((e) =>
-                    panelsRemoved.push({
-                        tgId: e.tabGroup.id,
-                        panelId: e.panelId,
-                    })
-                ),
-                dockview.api.onDidTabGroupChange((e) =>
-                    changes.push(e.tabGroup.id)
-                )
-            );
-
-            // Create a new tab group and verify events fire
-            const newTg = dockview.api.createTabGroup({
-                groupId: restoredGroup.id,
-                label: 'New',
-                color: 'purple',
-            });
-            expect(created).toEqual([newTg.id]);
-
-            // Add a panel to the new group (panel1 is already in the
-            // restored tab group, so it will be removed from there first)
-            dockview.api.addPanelToTabGroup({
-                groupId: restoredGroup.id,
-                tabGroupId: newTg.id,
-                panelId: 'panel1',
-            });
-            // One removal from the restored group, one addition to newTg
-            expect(panelsRemoved.length).toBe(1);
-            expect(panelsAdded.length).toBe(1);
-            expect(panelsAdded[0]).toEqual({
-                tgId: newTg.id,
-                panelId: 'panel1',
-            });
-
-            // Change the group label — should fire change event
-            newTg.setLabel('Updated');
-            expect(changes).toContain(newTg.id);
-
-            // Remove the panel
-            panelsRemoved.length = 0;
-            dockview.api.removePanelFromTabGroup({
-                groupId: restoredGroup.id,
-                panelId: 'panel1',
-            });
-            expect(panelsRemoved.length).toBe(1);
-            expect(panelsRemoved[0]).toEqual({
-                tgId: newTg.id,
-                panelId: 'panel1',
-            });
-
-            // Tab group should auto-destroy when empty
-            expect(destroyed).toContain(newTg.id);
-
-            disposable.dispose();
-        });
-
-        test('tab group events fire for all lifecycle operations', () => {
-            dockview.layout(1000, 1000);
-
-            const panel1 = dockview.addPanel({
-                id: 'panel1',
-                component: 'default',
-            });
-            const panel2 = dockview.addPanel({
-                id: 'panel2',
-                component: 'default',
-            });
-            const panel3 = dockview.addPanel({
-                id: 'panel3',
-                component: 'default',
-            });
-
-            const groupId = panel1.group.id;
-
-            const created: string[] = [];
-            const destroyed: string[] = [];
-            const panelsAdded: { tgId: string; panelId: string }[] = [];
-            const panelsRemoved: { tgId: string; panelId: string }[] = [];
-            const changes: string[] = [];
-
-            const disposable = new CompositeDisposable(
-                dockview.api.onDidCreateTabGroup((e) =>
-                    created.push(e.tabGroup.id)
-                ),
-                dockview.api.onDidDestroyTabGroup((e) =>
-                    destroyed.push(e.tabGroup.id)
-                ),
-                dockview.api.onDidAddPanelToTabGroup((e) =>
-                    panelsAdded.push({
-                        tgId: e.tabGroup.id,
-                        panelId: e.panelId,
-                    })
-                ),
-                dockview.api.onDidRemovePanelFromTabGroup((e) =>
-                    panelsRemoved.push({
-                        tgId: e.tabGroup.id,
-                        panelId: e.panelId,
-                    })
-                ),
-                dockview.api.onDidTabGroupChange((e) =>
-                    changes.push(e.tabGroup.id)
-                )
-            );
-
-            // 1. Create
-            const tg = dockview.api.createTabGroup({
-                groupId,
-                label: 'Test',
-                color: 'blue',
-            });
-            expect(created).toEqual([tg.id]);
-
-            // 2. Add panels
-            dockview.api.addPanelToTabGroup({
-                groupId,
-                tabGroupId: tg.id,
-                panelId: 'panel1',
-            });
-            dockview.api.addPanelToTabGroup({
-                groupId,
-                tabGroupId: tg.id,
-                panelId: 'panel2',
-            });
-            dockview.api.addPanelToTabGroup({
-                groupId,
-                tabGroupId: tg.id,
-                panelId: 'panel3',
-            });
-            expect(panelsAdded).toEqual([
-                { tgId: tg.id, panelId: 'panel1' },
-                { tgId: tg.id, panelId: 'panel2' },
-                { tgId: tg.id, panelId: 'panel3' },
-            ]);
-
-            // 3. Change label and color
-            changes.length = 0;
-            tg.setLabel('Updated');
-            tg.setColor('red');
-            expect(changes.length).toBe(2);
-
-            // 4. Collapse and expand
-            tg.collapse();
-            tg.expand();
-
-            // 5. Remove one panel
-            dockview.api.removePanelFromTabGroup({
-                groupId,
-                panelId: 'panel2',
-            });
-            expect(panelsRemoved).toEqual([{ tgId: tg.id, panelId: 'panel2' }]);
-
-            // 6. Dissolve — should remove remaining panels and destroy
-            panelsRemoved.length = 0;
-            dockview.api.dissolveTabGroup({
-                groupId,
-                tabGroupId: tg.id,
-            });
-
-            expect(panelsRemoved).toEqual(
-                expect.arrayContaining([
-                    { tgId: tg.id, panelId: 'panel1' },
-                    { tgId: tg.id, panelId: 'panel3' },
-                ])
-            );
-            expect(destroyed).toEqual([tg.id]);
-
-            disposable.dispose();
-        });
     });
 
     describe('edge panels', () => {
@@ -12100,72 +11725,6 @@ describe('dockviewComponent', () => {
             expect(dv.getEdgeGroup('left')).toBeDefined();
             dv.dispose();
         });
-    });
-
-    test('regression: tab-group events fire on a source group born during moveGroup-from-edge', () => {
-        // moveGroup from an edge group to the grid creates a fresh `source`
-        // group via this.createGroup() and adds it via gridview.addView(...) —
-        // bypassing doAddGroup, so BaseGrid._onDidAdd never fires for source.
-        // TabGroupChipsModule attaches per-group event forwarding via
-        // host.onDidAddGroup (driven by BaseGrid._onDidAdd through the
-        // gated callback in DockviewComponent); since neither path fires,
-        // attachToGroup never runs and subsequent api.createTabGroup on
-        // the moved-source group doesn't propagate to dockview.api.onDidCreateTabGroup.
-        function createFixedDockview(
-            c: HTMLElement,
-            positions: ('top' | 'bottom' | 'left' | 'right')[]
-        ): DockviewComponent {
-            const dv = new DockviewComponent(c, {
-                createComponent(options) {
-                    return new PanelContentPartTest(options.id, options.name);
-                },
-            });
-            for (const position of positions) {
-                dv.addEdgeGroup(position, { id: `${position}-group` });
-            }
-            return dv;
-        }
-
-        const c = document.createElement('div');
-        const dv = createFixedDockview(c, ['left']);
-        dv.layout(1000, 800);
-
-        const edgeGroup = dv.groups.find(
-            (g) => g.api.location.type === 'edge'
-        )!;
-
-        const target = dv.addPanel({ id: 'center', component: 'default' });
-        dv.addPanel({
-            id: 'edge-p1',
-            component: 'default',
-            position: { referenceGroup: edgeGroup.id },
-        });
-
-        // Move the edge group to the grid — internally creates a fresh
-        // `source` group via createGroup + gridview.addView (the path that
-        // bypasses doAddGroup / _onDidAddGroup).
-        dv.moveGroup({
-            from: { group: edgeGroup },
-            to: { group: target.api.group, position: 'right' },
-        });
-
-        const movedPanel = dv.getGroupPanel('edge-p1')!;
-        const movedGroup = movedPanel.group;
-
-        // Subscribe AFTER the move so we can prove the event propagates.
-        const events: string[] = [];
-        dv.api.onDidCreateTabGroup((e) => events.push(e.tabGroup.id));
-
-        // Trigger a tab-group operation on the moved-source group.
-        const tg = dv.api.createTabGroup({
-            groupId: movedGroup.id,
-            label: 'test',
-        });
-
-        // Should fire — but doesn't because attachToGroup never ran.
-        expect(events).toEqual([tg.id]);
-
-        dv.dispose();
     });
 
     test('regression: HeaderActions renderer disposed when group removed during a move (movingLock leak)', () => {
