+++ packages/dockview-core/src/dockview/modules.ts
@@ -2,10 +2,11 @@
  * Internal module system for dockview.
  *
  * Modules are feature bundles that register services into the dockview
- * component. This is currently an internal-only system used to refactor
- * the core component into independently-extracted features. The public
- * surface (a `modules` option, framework wrappers, opt-in registration)
- * is reserved for a future major version.
+ * component. `registerModules(...)` is the one public entry point — it lets a
+ * sibling package contribute modules that `DockviewComponent` picks up at
+ * construction. The richer opt-in surface (a per-component `modules` option,
+ * framework wrappers) is still reserved for a future version; the module
+ * authoring API (`defineModule`, the service contracts) remains internal.
  */
 
 import { IDisposable } from '../lifecycle';
@@ -195,4 +196,42 @@ export class ModuleRegistry<THost> implements IDisposable {
             }
         }
     }
+}
+
+/**
+ * Process-global list of modules registered via {@link registerModules}.
+ * `DockviewComponent` appends these to its built-in set at construction, so
+ * importing a package that calls `registerModules(...)` (e.g. `dockview`)
+ * makes those modules available to every component in the process —
+ * modelled on AG Grid's `ModuleRegistry.registerModules`.
+ */
+const _globalModules: DockviewModule<any>[] = [];
+
+/**
+ * Register modules globally. Idempotent per `moduleName` — registering the
+ * same module twice is a no-op. Intended to be called once at import time by
+ * the package that bundles a given set of modules.
+ */
+export function registerModules(modules: DockviewModule<any>[]): void {
+    for (const module of modules) {
+        if (_globalModules.some((m) => m.moduleName === module.moduleName)) {
+            continue;
+        }
+        _globalModules.push(module);
+    }
+}
+
+/**
+ * Returns the globally-registered modules (a copy). `DockviewComponent` reads
+ * this to extend its built-in module set.
+ */
+export function getRegisteredModules(): DockviewModule<any>[] {
+    return [..._globalModules];
+}
+
+/**
+ * For tests — clears the global module registry.
+ */
+export function clearRegisteredModules(): void {
+    _globalModules.length = 0;
 }
