+++ packages/dockview-modules/tsconfig.esm.json
@@ -1,0 +1,13 @@
+{
+    "extends": "../../tsconfig.base.json",
+    "compilerOptions": {
+        "module": "ES2020",
+        "moduleResolution": "node",
+        "target": "es6",
+        "outDir": "dist/esm",
+        "tsBuildInfoFile": ".build/tsconfig.tsbuildinfo.esm",
+        "rootDir": "src"
+    },
+    "include": ["src"],
+    "exclude": ["**/node_modules", "src/__tests__"]
+}
