+++ tsconfig.json
@@ -8,6 +8,9 @@
             "path": "./packages/dockview-core"
         },
         {
+            "path": "./packages/dockview-modules"
+        },
+        {
             "path": "./packages/dockview-vue"
         },
         {
