+++ tsconfig.test.json
@@ -5,7 +5,8 @@
         "paths": {
             "dockview-core": ["packages/dockview-core/src/index.ts"],
             "dockview": ["packages/dockview/src/index.ts"],
-            "dockview-react": ["packages/dockview-react/src/index.ts"]
+            "dockview-react": ["packages/dockview-react/src/index.ts"],
+            "dockview-modules": ["packages/dockview-modules/src/index.ts"]
         },
         "jsx": "react-jsx",
         "noImplicitAny": false,
